+++ src/core_unix.mli
@@ -693,6 +693,50 @@ val flock : File_descr.t -> Flock_command.t -> bool
     console window, [false] otherwise. *)
 val isatty : File_descr.t -> bool
 
+(** {6 Mapping files into memory} *)
+
+(** Memory mapping of a file as a big array.
+    [map_file fd kind layout ~shared dims] returns a big array of kind [kind],
+    layout [layout], and dimensions as specified in [dims].
+
+    The data contained in this big array are the contents of the file referred to by the
+    file descriptor [fd].
+
+    The optional [pos] parameter is the byte offset in the file of the data being mapped.
+    It defaults to 0.
+
+    If [shared] is [true], all modifications performed on the array are reflected in
+    the file.  This requires that [fd] be opened with write permissions.  If [shared]
+    is [false], modifications performed on the array are done in memory only, using
+    copy-on-write of the modified pages; the underlying file is not affected.
+
+    To adjust automatically the dimensions of the big array to the actual size of the
+    file, the major dimension (that is, the first dimension for an array with C layout,
+    and the last dimension for an array with Fortran layout) can be given as
+    [-1]. [map_file] then determines the major dimension from the size of the file.
+    The file must contain an integral number of sub-arrays as determined by the non-major
+    dimensions, otherwise [Failure] is raised. If all dimensions of the big array are
+    given, the file size is matched against the size of the big array.
+    If the file is larger than the big array, only the initial portion of the file is
+    mapped to the big array.  If the file is smaller than the bigarray, the file is
+    automatically grown to the size of the big array. This requires write permissions on
+    [fd].
+
+    Array accesses are bounds-checked, but the bounds are determined by the initial call
+    to [map_file]. Therefore, you should make sure no other process modifies the mapped
+    file while you're accessing it, or a SIGBUS signal may be raised. This happens,
+    for instance, if the file is shrunk. [Invalid_argument] or [Failure] may be raised
+    in cases where argument validation fails.
+    @since 4.05.0 *)
+val map_file
+  : File_descr.t
+  -> ?pos:int64
+  -> ('a, 'b) Bigarray.kind
+  -> 'c Bigarray.layout
+  -> shared:bool
+  -> int array
+  -> ('a, 'b, 'c) Bigarray.Genarray.t
+
 (** {6 Operations on file names} *)